refactor: move DatePicker to its own component
